Fix bug in quantize (rounding).
Turns out that automatically stripping trailing zeroes after the radix
mark was causing problems for my rounding.  I needed the value to have
the same exponent after adding the roundoff value as it did before, and
the automatic removal of trailing zeroes in clean_value() and make()
were booching that.

I decided that stripping the trailing zeroes is not actually all that
important, especially since the output formatters will do that for you
and there's always compress() if you really need to get rid of them.