Add tests for ->inverse().
Also, adjust the default scale of ->inverse() to be more useful in
typical cases.  The scale of the decimal was not a good choice, since
in general the larger the value, the more precision is needed to
adequately represent its inverse.  I settled on a default scale of
exponent + 1, if the exponent is non-negative, otherwise zero.